backend/requirements.txt
pyramid==2.0.2
pyramid-cors==1.3
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
PyJWT==2.8.1
bcrypt==4.1.1
python-dotenv==1.0.0
alembic==1.13.0
